Minor fixes and tidying some code: - Fixed crash occurring when using EFFECTLOCATION with invalid SRC; - Increased MAX_EXTCMD_ARG_LEN from 30 to 40 to allow bigger arguments while using a custom packet (Issue #280); - Moved cross-platform macros for string functions from common.h/cpp to sstring.h/cpp; - Moved some calls to CLanguageID methods from sphereproto.h to common.cpp; - Replaced references to ITOA with Str_FromI; - Changed the order of the parameters in Str_From*, to match the one of ITOA and LTOA; - Removed LTOA (unused) macro and replaced ATOI macro with plain "atoi"; - Renamed STRREV to Str_Reverse; - Added missing zero-initialization in CRect constructor; - Avoided redundant code in CScriptObj::r_LoadVal, SSC_VAR/0; - Minor optimization in CClient::r_LoadVal, while checking for the CTAG keyword; - Fixed incorrect parsing of quoted arguments to ACCOUNT.x.TAG; - Changed default value of argument fDeleteZero from false to true in any SetNum, ModNum and similar methods (before they were mixed); - Added cast to int64 for *_PER_SEC macros, to avoid integer overflows.

+++ src/game/CServerTime.h
@@ -15,9 +15,9 @@
 */
 struct CServerTime
 {
-	#define TICKS_PER_SEC   10      // Amount of ticks to advance in a second.
-    #define TENTHS_PER_SEC  10      // Tenths in a second (backwards).
-    #define MSECS_PER_SEC   1000    // Milliseconds in a second (to avoid magic numbers).
+	#define TICKS_PER_SEC   (int64)10      // Amount of ticks to advance in a second.
+    #define TENTHS_PER_SEC  (int64)10      // Tenths in a second (backwards).
+    #define MSECS_PER_SEC   (int64)1000    // Milliseconds in a second (to avoid magic numbers).
 
     #define MSECS_PER_TENTH (int64)100
     #define MSECS_PER_TICK  (int64)(MSECS_PER_SEC / TICKS_PER_SEC) // Milliseconds lapse between one tick and another.
